test(core): move env-agnostic tests into @flareapp/core
Moves 16 test files and the helpers directory from packages/js/tests/ to
packages/core/tests/. Adds vitest.config.ts with node environment (per-file
jsdom overrides remain). Adds jsdom devDependency. Drops --passWithNoTests
now that tests exist. Updates import paths in createStackTrace.test.ts,
fileReader.test.ts, and redactUrl.test.ts. Renames redactFullPath to
redactUrlQuery throughout the test. Re-creates packages/js/tests/helpers/ as
a self-contained copy for the remaining browser-specific js tests.
