+++ packages/core/package.json
@@ -33,7 +33,7 @@
     "scripts": {
         "prepublishOnly": "npm run build",
         "build": "tsdown src/index.ts --format cjs,esm --dts --env.FLARE_JS_CLIENT_VERSION=$(node -p \"require('./package.json').version\") --clean",
-        "test": "vitest run --passWithNoTests",
+        "test": "vitest run",
         "typescript": "tsc --noEmit",
         "release": "release-it"
     },
@@ -41,6 +41,7 @@
         "error-stack-parser": "^2.0.2"
     },
     "devDependencies": {
+        "jsdom": "^26.1.0",
         "tsdown": "^0.20.3",
         "typescript": "^5.7.0",
         "vitest": "^4.0.18"
