+++ packages/core/tests/configure.test.ts
