+++ packages/core/tests/convertToError.test.ts
