+++ packages/core/tests/glowsToEvents.test.ts
