+++ packages/core/tests/hooks.test.ts
