+++ packages/core/tests/sampleRate.test.ts
