+++ packages/core/tests/setEntryPoint.test.ts
