+++ packages/core/tests/setSdkInfo.test.ts
